Anchor design-system-v3 branch on master and fix Filament admin styling
Records the current working state on top of master (Filament admin resources, pages, widgets, settings; frontend and email-template updates) and fixes the broken admin styling/asset pipeline:

- Move the custom panel theme from a brittle ->assets(Css::make()->html(app(Vite)())) registration to a STYLES_AFTER render hook loading the compiled theme via @vite. This fixes CSS cascade order (core styles first, overrides after), makes the theme SPA-safe (data-navigate-track), and stops 'php artisan filament:assets' from crashing.

- Un-ignore public/build in .gitignore so the precompiled Vite assets are committed and deployed to the Node-less production server (the generic 'build/' rule was silently excluding them).

+++ app/Http/Controllers/Frontend/CartController.php
@@ -169,8 +169,8 @@ class CartController extends Controller
             return [
                 'id' => $item->id,
                 'quantity' => $item->quantity,
-                'price' => (float) $product->price,
-                'line_total' => (float) bcmul((string) $product->price, (string) $item->quantity, 2),
+                'price' => $product->price,
+                'line_total' => bcmul((string) $product->price, (string) $item->quantity, 2),
                 'oem_number' => $product->oem_number,
                 'name' => trans_field($product->name),
                 'condition' => $product->condition,
